clone sessions by hardlinking (not copying)

+++ simserver/simserver.py
@@ -48,6 +48,20 @@ SHARD_SIZE = 65536 # spill index shards to disk in SHARD_SIZE-ed chunks of docum
 
 JOURNAL_MODE = 'OFF' # don't keep journals in sqlite dbs
 
+HARDLINK_SESSIONS = True # when cloning an index, use hardlink for better performance (only works on UNIX systems)
+if HARDLINK_SESSIONS:
+    import __builtin__
+    _open = __builtin__.open
+
+    def my_open(name, mode='r', *args, **kwargs):
+        """Simulate copy-on-write, by deleting the file first if it exists"""
+        if 'w' in mode and os.path.exists(name):
+            os.remove(name)
+        return _open(name, mode, *args, **kwargs)
+
+    __builtin__.open = my_open
+
+
 def merge_sims(oldsims, newsims, clip=TOP_SIMS):
     """Merge two precomputed similarity lists, truncating the result to `clip` most similar items."""
     if oldsims is None:
@@ -824,13 +838,17 @@ class SessionServer(gensim.utils.SaveLoad):
 
         logger.info("opening a new session")
         logger.info("removing %s" % self.loc_session)
-        try:
+        if os.path.isdir(self.loc_session):
             shutil.rmtree(self.loc_session)
-        except:
-            logger.info("failed to delete %s" % self.loc_session)
         logger.info("cloning server from %s to %s" %
                     (self.loc_stable, self.loc_session))
-        shutil.copytree(self.loc_stable, self.loc_session)
+        if HARDLINK_SESSIONS:
+            # on unix systems, copy by hardlinking files
+            # requires overriding built-in open('w') -- see `my_open` above
+            gensim.utils.copytree_hardlink(self.loc_stable, self.loc_session)
+        else:
+            # (much) less efficient than hardlinking, but safer
+            shutil.copytree(self.loc_stable, self.loc_session)
         self.session = SimServer(self.loc_session, use_locks=self.use_locks)
         #self.lock_update.acquire() # no other thread can call any modification methods until commit/rollback
